finish combat system

+++ Assets/Game/Scripts/Interfaces/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(float damage);
+    void Damage(int damage);
 }
